Extracted ContractWrapper parent class
