+++ src/actions/deploy_actions.js
@@ -57,10 +57,6 @@ export default class DeployActions {
     return this.deployer.deploy(contractsToDeploy, {...genesisContracts, ...recycledContracts}, [], {multiplexer: {owner: this.sender}});
   }
 
-  async setMultiplexerAsAnOwner() {
-    return this.deployer.setMultiplexerAsAnOwner(contractJsons);
-  }
-
   async recycleStorageContracts() {
     const storageContractNames = this.headWrapper.availableStorageCatalogueContracts;
     const recycled = {};
